+++ src/main/java/com/likelion/zooting/domain/match/policy/MatchPolicy.java
@@ -10,7 +10,7 @@ import java.util.Map;
  * <p><b>[매칭 원칙]</b></p>
  * <ul>
  * <li><b>최우선 순위:</b> 합산 점수가 높은 쌍부터 우선적으로 매칭을 확정합니다.</li>
- * <li><b>필터링 조건:</b> 동물상, 관심사, 영화 장르 중 최소 하나라도 겹치지 않아 합산 점수가 0점인 경우 매칭 대상에서 제외합니다.</li>
+ * <li><b>필터링 조건:</b> 총 매칭 점수가 90점 이상인 것만 통과시킵니다.</li>
  * </ul>
  *
  * <p><b>[동점 발생 시 우선순위 처리]</b></p>
@@ -26,7 +26,7 @@ public interface MatchPolicy {
      *
      * <p><b>[동작 과정]</b></p>
      * <ol>
-     * <li><b>필터링:</b> 전체 점수판에서 0점을 초과하는 유효한 결과만을 선별합니다.</li>
+     * <li><b>필터링:</b> 전체 점수판에서 90점을 초과하는 유효한 결과만을 선별합니다.</li>
      * <li><b>정렬:</b> 선별된 매칭 후보들을 점수가 높은 순(내림차순)으로 정렬합니다.</li>
      * <li><b>확정:</b> 정렬된 순서대로 매칭 그룹에 포함시키며, 매칭 쌍 중 한 명이라도 이미 매칭이 완료된 상태라면 해당 쌍은 제외합니다.</li>
      * </ol>
